requirements.txt
coloredlogs==15.0.1
coverage==6.0.2
pytest-cov==3.0.0
pytest~=6.2.5
python-dotenv==0.20.0
requests==2.22.0
pre-commit==2.20.0
types-requests==2.28.11.2
